Renamed Icons.tsx to LegacyIcons.tsx
since shall be replaced with needle icons and to not confuse with the other Icons.tsx file.
