+++ src/browser/components/FileDrop/FileDrop.tsx
@@ -21,7 +21,7 @@ import React, { useState } from 'react'
 import { connect } from 'react-redux'
 import { withBus } from 'react-suber'
 
-import { FileDropIcon } from 'browser-components/icons/Icons'
+import { FileDropIcon } from '../icons/LegacyIcons'
 
 import arrayHasItems from '../../../shared/utils/array-has-items'
 import {
