+++ src/browser/components/SavedScripts/SavedScripts.tsx
@@ -22,7 +22,7 @@ import React, { useState } from 'react'
 import { useDrop } from 'react-dnd'
 import uuid from 'uuid'
 
-import { AddIcon } from 'browser-components/icons/Icons'
+import { AddIcon } from '../icons/LegacyIcons'
 
 import {
   ExportButton,
