+++ src/browser/components/icons/LegacyIcons.tsx
