+++ src/browser/modules/DBMSInfo/styled.tsx
@@ -20,7 +20,7 @@
 import React from 'react'
 import styled from 'styled-components'
 
-import { PlainPlayIcon } from 'browser-components/icons/Icons'
+import { PlainPlayIcon } from 'browser-components/icons/LegacyIcons'
 
 export const StyledTable = styled.table``
 export const StyledKey = styled.td`
