+++ src/browser/modules/Sidebar/Sidebar.tsx
@@ -29,7 +29,7 @@ import {
   GuideDrawerIcon,
   ProjectFilesIcon,
   SettingsIcon
-} from 'browser-components/icons/Icons'
+} from 'browser-components/icons/LegacyIcons'
 
 import DatabaseDrawer from '../DBMSInfo/DBMSInfo'
 import BrowserSync from '../Sync/BrowserSync'
