+++ src/browser/modules/Stream/Auth/revealable-password-input.tsx
@@ -16,7 +16,7 @@
  */
 import React, { useEffect, useRef } from 'react'
 
-import { EyeIcon, EyeSlashIcon } from 'browser-components/icons/Icons'
+import { EyeIcon, EyeSlashIcon } from 'browser-components/icons/LegacyIcons'
 
 import {
   StyledConnectionTextInput,
