+++ src/browser/modules/User/UserAdd.tsx
@@ -23,7 +23,7 @@ import { connect } from 'react-redux'
 import { withBus } from 'react-suber'
 import uuid from 'uuid'
 
-import { CloseIcon } from 'browser-components/icons/Icons'
+import { CloseIcon } from 'browser-components/icons/LegacyIcons'
 
 import RolesSelector from './RolesSelector'
 import { StyleRolesContainer, StyledInput } from './styled'
